Refactor extrinsic handling, #179
