+++ backend/packages/block-scan/src/scan/extrinsic/normalize.js
@@ -1,5 +1,5 @@
 const {
-  utils: { extractExtrinsicEvents, isExtrinsicSuccess },
+  utils: { isExtrinsicSuccess },
   call: { normalizeCall },
 } = require("@osn/scan-common");
 
@@ -59,20 +59,6 @@ function normalizeExtrinsic(extrinsic, events, indexer) {
   return obj;
 }
 
-function normalizeExtrinsics(extrinsics = [], blockEvents = [], blockIndexer) {
-  let index = 0;
-  let normalizedExtrinsics = [];
-  for (const extrinsic of extrinsics) {
-    const events = extractExtrinsicEvents(blockEvents, index);
-    const extrinsicIndexer = { ...blockIndexer, extrinsicIndex: index++ };
-
-    const normalized = normalizeExtrinsic(extrinsic, events, extrinsicIndexer);
-    normalizedExtrinsics.push(normalized);
-  }
-
-  return normalizedExtrinsics;
-}
-
 module.exports = {
-  normalizeExtrinsics,
+  normalizeExtrinsic,
 };
